pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-collector-unit-poc"
dynamic = ["version"]
description = ""
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = []
authors = [
  { name = "Alexander Kupreyeu" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "numpy>=2.1.0",
  "pandas>=2.2.2",
  "scikit-learn>=1.5.1",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "more-itertools>=10.5.0",
  "SQLAlchemy>2.0.0",
  "alembic",
  "aiosqlite"
]

[project.optional-dependencies]
dev = ["ruff", "basedpyright"]
test = ["pytest"]
notebook = [
  "jupyter",
  "tqdm>=4.66.5",
]
orchestrator = [
  "apscheduler",
]
nicegui = [
  "nicegui"
]
server = [
  "fastapi[standard]",
  "fastapi-cors",
  "fastapi-users[sqlalchemy]",
  "uvicorn",
  "sqladmin[full]",
  "boto3"
]

[project.urls]
Documentation = "https://github.com/AlexKupreev/data-collector-unit-poc#readme"
Issues = "https://github.com/AlexKupreev/data-collector-unit-poc/issues"
Source = "https://github.com/AlexKupreev/data-collector-unit-poc"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "src/data_collector_unit_poc/__about__.py"

[tool.hatch.envs.default]
pre-install-commands = ["echo 'start env installation'"]
post-install-commands = ["echo 'end env installation'"]
python = "3.11"
installer = "uv"
type = "virtual"
path = ".venv"
features = ["dev", "test", "notebook", "server", "nicegui", "orchestrator"]
[tool.hatch.envs.default.scripts]
jobs = "dagster dev -h 0.0.0.0 -p 3000"

[tool.dagster]
module = "dasgter"

[tool.hatch.envs.prod]
features = ["server", "nicegui", "orchestrator"]

[tool.hatch.envs.web]
path = ".venv_web"
features = ["dev", "test", "nicegui"]

[tool.hatch.envs.test]
path = ".venv_test"
features = ["test"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/data_collector_unit_poc tests}"

[tool.coverage.run]
source_pkgs = ["data_collector_unit_poc", "tests"]
branch = true
parallel = true
omit = [
  "src/data_collector_unit_poc/__about__.py",
]

[tool.coverage.paths]
dagster_databox = ["src/data_collector_unit_poc", "*/data-collector-unit-poc/src/data_collector_unit_poc"]
tests = ["tests", "*/data-collector-unit-poc/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
